web/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
aiosqlite==0.20.0
pytz==2024.1
python-dotenv==1.0.1
